requirements.txt
fastapi
uvicorn
langgraph
httpx
python-dotenv
pydantic
huggingface_hub
